// IMFS-Web-FrontEnd/src/app/models/application/application.model.ts
export class ApplicationSearchModel {
    applicationNumber?: number;
    status: number;
    financeType?: string[];
    fromDate?: Date;
    toDate?: Date;
    endCustomerName: string;
    TriggerSource?:string;
    resellerId:number;
}


export class ApplicationSearchResponseModel {
    id: number;
    applicationNumber: number;
    status: number;
    statusDescr: string;
    endCustomerName: string;
    resellerId: string;
    financeTotal: number;
    quoteTotal: number;
    financeType: number;
    financeTypeName: string;
    createdDate: Date;
    displayLabel: string;
}